requirements.txt
autopep8==1.5.7
Django==3.2.6
django-render-partial==0.4
djongo==1.3.6
jalaali==0.0.1
jdatetime==3.6.2
Pillow==8.3.1
pymongo==3.12.0
xhtml2pdf==0.2.5
openpyxl==3.0.9
paho-mqtt==1.6.1
plotly==5.4.0
pandas==1.3.4
kaleido==0.2.1